Allow skipping changing directory to tmpdir when using packs/rspec/support

+++ lib/packs/rspec/support.rb
@@ -10,12 +10,18 @@ RSpec.configure do |config|
 
   # Eventually, we could make this opt-in via metadata so someone can use this support without affecting all their tests.
   config.around do |example|
-    prefix = [File.basename($0), Process.pid].join('-') # rubocop:disable Style/SpecialGlobalVars
-    tmpdir = Dir.mktmpdir(prefix)
-    Dir.chdir(tmpdir) do
+    if example.metadata[:skip_chdir_to_tmpdir]
       example.run
+    else
+      begin
+        prefix = [File.basename($0), Process.pid].join('-') # rubocop:disable Style/SpecialGlobalVars
+        tmpdir = Dir.mktmpdir(prefix)
+        Dir.chdir(tmpdir) do
+          example.run
+        end
+      ensure
+        FileUtils.rm_rf(tmpdir)
+      end
     end
-  ensure
-    FileUtils.rm_rf(tmpdir)
   end
 end
